Split the bar label adapters out of the placement engine
labelPlacement.ts held both the placement engine and the bar-family
adapter block that feeds it. The dependency runs one way, so the adapters
move to barLabelGeometry.ts and the engine drops from 2874 to 2270 lines.
main.ts re-exports both, so no consumer import changes.

Alongside the move:

- boundByRegion becomes a required positive flag set by both
  LabelFitDescriptor factories, replacing a field that was only ever
  `false` or absent and read as `=== false`.
- The engine-private insideMarkerContainer becomes
  insideMarkerCandidateContainer, no longer colliding with the community
  export of the same name that answers a different question.
- shrinkCompassCandidate drops from twelve parameters to four; the other
  eight were cascade scratch the caller had just written.
- Both candidate loops share one flushCandidateBox, mirroring the existing
  slideShrunkCandidate/shrunkOffset pair.
- placedBarLabelTargets, which had no consumers, is inlined into
  applyPlacedBarLabelVisibility.
